fix: use ordering of PRs ignoring bad ones
